Dropped dbupgrade files used for testing PK/FK migrations. Added dbmigrations.ts to keep all db.version().upgrade() logic isolated from db.js.
